+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## next
 
+* Support LLVM 23:
+  * `DIBasicType'` now has additional `dibt{Scope,File,Line}` fields.
+
 ## 0.15.0.0 -- 2026-08-27
 
 * Add `LLVM.Combine` module with `llvmModuleCombine` function.  This is the